LVGL_HelloWorld * tft: touch_read_data I have to update this function, as the expected method was not working, withe the commented function, data was not transmitted properly

+++ ESP-IDF/LVGL_HelloWorld/main/tft.h
@@ -48,6 +48,6 @@ void tft_init( void );
 void tft_delay_ms(uint32_t delay);
 void tft_send_cmd( uint8_t cmd, const uint8_t *data, int len );
 void tft_send_data( const uint8_t *data, int len );
-void touch_read_data( uint8_t cmd, uint8_t *data, int len );
+void touch_read_data( uint8_t cmd, uint8_t *data, uint8_t len );
 
 #endif /* MAIN_TFT_H_ */
